fix display of an erroneous selection when switching tabs
scenario (before the fix):
- open 2 songs
- in song 1, click in measure 4, without click&drag : no selection
- switch to song 2
- switch back to song 1
a selection appears in measure 4 where the cursor was

Explanation: introduced in 0d728c2ab26b2256e5ac00d7124182912bd3370f when
fixing one bug backported from 2.0beta.
In 2.0beta, when user switches song, a selection is initialized where the
cursor was.
BUT, since a selection of a single beat does not work, nothing happens

once the possibility to select 1 single beat selection was reintroduced, the
consequences became visible

fix: when switching song, save/restore selection only if it's not empty