remove links to external web pages from help, authors updated
keeping the URLs as text, but removing  "<a>" tag

trade-off for security, now that doc is opened by embedded browser.
A click on such a link and the embedded browser is in the wild, connected to Internet.
If a vulnerability is discovered for embedded browser, it could have a severe
impact on TuxGuitar's users, and most probably it would take a very long
time before an eventual patch reaches all users of a TuxGuitar release.
Since there is no control to enter a URL, this limits de facto the browser
to the delivered files in local file-system, reducing the attack surface.

authors updated: that's for self-satisfaction ;)