workaround for an external bug
in SWT configuration, with some versions of Linux/KDE, estimated width of
checkboxes and radio buttons by SWT is underestimated
This results in truncated strings shown to user

This is NOT a fix, just a workaround.
To be removed one day when external issue is fixed

Consequence: some controls may have an extra (useless) margin of 10px on
other platforms (Windows, mac, freebsd).
I tried to add a parameter, so that the workaround would only apply to
Linux version, but this would have required to add some dependencies
between modules in pom files, I think it's not worth it.


see https://github.com/helge17/tuxguitar/issues/52