bend and multitrack: better horizontal spacing
this is a trade-off
restoring vertical alignment of beats in multitrack mode when bends
are present is complex. And this should lead to a significant waste of space
So, this alignment is not strictly respected, especially when many wide
bend movements are displayed in one track and none in the other(s).
Max width of bend movements of all tracks is "distributed" linearly in each
measure of each track.
At least, available horizontal space is used in a more optimal way.

see https://github.com/helge17/tuxguitar/issues/242