Improve tuplet handling (#331)
Improvements related to issue #118

This should allow tuxguitar to handle tuplets a lot better.
Please note that because the value of TGDuration.QUARTER_TIME is 960, tuxguitar is not able to compute precise durations for 7, 9, 11 and 13 tuplets. Idealy TGDuration.QUARTER_TIME should be set to 2882880 in order for tuxguitar to compute precise durations for all these tuplets, but simply changing TGDuration.QUARTER_TIME currently has some unintended consequences that would need to be handled first.
Because of this precision lose I have also added some patches that would fix some basic rounding errors when TGDuration.QUARTER_TIME is 960.