Fix Move one Beat left button
Addresses #585 .

Original issue: When moving to the left, the `theMove` parameter is negative, and would go down the path of removing beats. With this, we would end up removing the beats that had the notes and would not re-add them.

Fix: Adds a hash-map of <time after move, list of notes pre-move> so we can keep track of what notes need to be moved, and add them to the new (post-move) beats.