fixes (partially) display of note names in scales on fretboard
At least, scales with flat base notes can be displayed (e.g. Eb),
and result should be ok for many common scales.
Not a perfect solution though.
It can't be correct in all cases, as there are no concepts of
double-sharp / double-flat in TuxGuitar currently.
E.g. D# major scale is displayed with 3 sharps (in an incorrect order)
instead of 5 sharps + 2 double-sharps.
Intentionally, E#, Fb, B# and Cb scales are not accessible.