refactoring
cleanup of 2ddade5c0921ae708af8929ad307c586ca696f9a for readability.
This was developed with the same logic as legacy versions (<2.0.0)
where preciseDuration was not available in TGDuration structure: working
with durations as fractions of time (homogeneous to 1/time: the bigger the
value, the shorter the note)

It's now possible to take benefit from this precise data. Working with
time instead of inverse of time makes code more understandable